feat(04-08): implement scheduled campaign dispatch + sender-ID state machine (GREEN)
- ScheduledCampaignDispatchJob: @Scheduled DB-poll dispatcher with testable dispatch(Instant) delegate; calls findByStatusAndScheduledAtBefore(SCHEDULED, now, page50), per-item try/catch, delegates to CampaignService.executeSend
- CampaignService.cancel(): IDOR-safe cancel from SCHEDULED/DRAFT→CANCELLED; idempotent
- CampaignController: POST /api/v1/campaigns/{id}/cancel endpoint
- SenderIdRepository, SenderIdService: request/approve/reject state machine with SenderIdDecided outbox event on decision
- SenderIdController: POST+GET /api/v1/sender-ids/requests with @Valid senderName max=11 alphanumeric
- SenderIdAdminController: POST /api/v1/internal/sender-ids/{id}/approve|reject (ROLE_ADMIN guarded)
- SecurityConfig: add JwtAuthenticationConverter mapping roles claim to Spring Security authorities
- Rule 1 fix: SenderIdIT.java type mismatch long vs int in isGreaterThan assertion (compile error)

+++ services/messaging-service/src/test/java/com/opendesk/messaging/SenderIdIT.java
@@ -192,7 +192,7 @@ class SenderIdIT extends AbstractMessagingIntegrationTest {
                 .filter(e -> e.getEventType().equals("SenderIdDecided"))
                 .toList();
 
-        assertThat(decidedEntries.size()).isGreaterThan(outboxCountBefore);
+        assertThat(decidedEntries.size()).isGreaterThan((int) outboxCountBefore);
 
         // Latest entry payload must contain REJECTED decision and senderName
         var latestDecided = decidedEntries.stream()
